refactor(language_server): split options for linting and formatting (#13627)
Splits the options intern, so we can define options for `ServerLinter` and `ServerFormatter`.
There is no option added. This comes later (maybe `experimental: bool`?)
