+++ crates/oxc_language_server/src/formatter/mod.rs
@@ -1,0 +1,1 @@
+pub mod options;
