+++ crates/oxc_language_server/src/formatter/options.rs
@@ -1,0 +1,5 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct FormatOptions;
