+++ crates/oxc_language_server/src/linter/mod.rs
@@ -1,5 +1,6 @@
 pub mod config_walker;
 pub mod error_with_position;
 pub mod isolated_lint_handler;
+pub mod options;
 pub mod server_linter;
 pub mod tsgo_linter;
